Rename AbstractRunner to Runner
