+++ docs/MavenSite/apidocs/de/viadee/bpm/vPAV/AbstractRunner.html
@@ -101,7 +101,7 @@ var activeTableTab = "activeTableTab";
 <li><a href="http://docs.oracle.com/javase/7/docs/api/java/lang/Object.html?is-external=true" title="class or interface in java.lang">java.lang.Object</a></li>
 <li>
 <ul class="inheritance">
-<li>de.viadee.bpm.vPAV.AbstractRunner</li>
+<li>de.viadee.bpm.vPAV.Runner</li>
 </ul>
 </li>
 </ul>
