+++ docs/MavenSite/apidocs/de/viadee/bpm/vPAV/class-use/AbstractRunner.html
@@ -4,7 +4,7 @@
 <head>
 <!-- Generated by javadoc (1.8.0_131) on Wed Sep 06 11:28:53 CEST 2017 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Uses of Class de.viadee.bpm.vPAV.AbstractRunner (viadee Process Application Validator - Maven Plugin 2.1.0-SNAPSHOT API)</title>
+<title>Uses of Class de.viadee.bpm.vPAV.Runner (viadee Process Application Validator - Maven Plugin 2.1.0-SNAPSHOT API)</title>
 <meta name="date" content="2017-09-06">
 <link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../../../script.js"></script>
@@ -71,7 +71,7 @@
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h2 title="Uses of Class de.viadee.bpm.vPAV.AbstractRunner" class="title">Uses of Class<br>de.viadee.bpm.vPAV.AbstractRunner</h2>
+<h2 title="Uses of Class de.viadee.bpm.vPAV.Runner" class="title">Uses of Class<br>de.viadee.bpm.vPAV.Runner</h2>
 </div>
 <div class="classUseContainer">
 <ul class="blockList">
