+++ src/main/java/de/viadee/bpm/vPAV/ProcessApplicationValidator.java
@@ -54,9 +54,9 @@ public class ProcessApplicationValidator {
         RuntimeConfig.getInstance().setApplicationContext(ctx);
         RuntimeConfig.getInstance().setBeanMapping(BeanMappingGenerator.generateBeanMappingFile(ctx));
         RuntimeConfig.getInstance().setClassLoader(ProcessApplicationValidator.class.getClassLoader());
-        AbstractRunner.viadeeProcessApplicationValidator();
+        Runner.viadeeProcessApplicationValidator();
 
-        return AbstractRunner.getfilteredIssues();
+        return Runner.getfilteredIssues();
     }
 
     /**
@@ -71,9 +71,9 @@ public class ProcessApplicationValidator {
         RuntimeConfig.getInstance().setApplicationContext(ctx);
         RuntimeConfig.getInstance().setBeanMapping(BeanMappingGenerator.generateBeanMappingFile(ctx));
         RuntimeConfig.getInstance().setClassLoader(ProcessApplicationValidator.class.getClassLoader());
-        AbstractRunner.viadeeProcessApplicationValidator();
+        Runner.viadeeProcessApplicationValidator();
 
-        return filterErrors(AbstractRunner.getfilteredIssues(), CriticalityEnum.ERROR);
+        return filterErrors(Runner.getfilteredIssues(), CriticalityEnum.ERROR);
     }
 
     /**
@@ -84,9 +84,9 @@ public class ProcessApplicationValidator {
     public static Collection<CheckerIssue> findModelInconsistencies() {
 
         RuntimeConfig.getInstance().setClassLoader(ProcessApplicationValidator.class.getClassLoader());
-        AbstractRunner.viadeeProcessApplicationValidator();
+        Runner.viadeeProcessApplicationValidator();
 
-        return AbstractRunner.getfilteredIssues();
+        return Runner.getfilteredIssues();
     }
 
     /**
@@ -97,9 +97,9 @@ public class ProcessApplicationValidator {
     public static Collection<CheckerIssue> findModelErrors() {
 
         RuntimeConfig.getInstance().setClassLoader(ProcessApplicationValidator.class.getClassLoader());
-        AbstractRunner.viadeeProcessApplicationValidator();
+        Runner.viadeeProcessApplicationValidator();
 
-        return filterErrors(AbstractRunner.getfilteredIssues(), CriticalityEnum.ERROR);
+        return filterErrors(Runner.getfilteredIssues(), CriticalityEnum.ERROR);
     }
 
     /**
