+++ src/test/java/de/viadee/bpm/vPAV/processing/checker/OverlapCheckerTest.java
@@ -42,6 +42,7 @@ import java.util.Collection;
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.xpath.XPathExpressionException;
 
+import de.viadee.bpm.vPAV.Runner;
 import org.camunda.bpm.model.bpmn.Bpmn;
 import org.camunda.bpm.model.bpmn.BpmnModelInstance;
 import org.camunda.bpm.model.bpmn.instance.BaseElement;
@@ -50,7 +51,6 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import org.xml.sax.SAXException;
 
-import de.viadee.bpm.vPAV.AbstractRunner;
 import de.viadee.bpm.vPAV.BpmnScanner;
 import de.viadee.bpm.vPAV.RuntimeConfig;
 import de.viadee.bpm.vPAV.config.model.Rule;
@@ -105,7 +105,7 @@ public class OverlapCheckerTest {
             issues.addAll(checker.check(element));
         }
 
-        AbstractRunner.resetSequenceFlowList();
+        Runner.resetSequenceFlowList();
 
         if (issues.size() > 0) {
             Assert.fail("correct model generates an issue");
@@ -139,7 +139,7 @@ public class OverlapCheckerTest {
             issues.addAll(checker.check(element));
         }
 
-        AbstractRunner.resetSequenceFlowList();
+        Runner.resetSequenceFlowList();
 
         if (issues.size() != 1) {
             Assert.fail("Incorrect model should generate an issue");
